Distance
Added a function that calculates the distance between the Banana Tree and the player and gives the player a banana if they are close enough. More work needs to be done to make sure the banana overlaps on top of the player and stays where the player is even when they move away from the tree.