coconut tree + coconut stall
added the coconut tree that gives the player a coconut when close and the coconut stall which takes the coconut from the player when close. Converted the tree classes to functions because they were unnecessary and rearranged the code so that it is easier to read. renamed some variables to better fit their use.